move docs stuff around
