+++ docs/src/routes/Footer.svelte
